Refactor: Remove Handlers project and added AspNetCore package
Refactored all code dependent on AspNetCore to a new `Infinity.Toolkit.AspNetCore` module, introducing ASP.NET Core-specific functionality. Updated project references and namespaces accordingly.

Key changes:
- Added `Infinity.Toolkit.AspNetCore` targeting .NET 9.0.
- Removed `Infinity.Toolkit.Handlers` project.
- Migrated `ToProblemDetails` method to `Infinity.Toolkit.AspNetCore`.
- Updated `HandlersSample.csproj` to use `Infinity.Toolkit.AspNetCore`.
- Adjusted target frameworks and dependencies across projects.
- Added documentation for `Infinity.Toolkit.AspNetCore` in `README.md`.

This refactor modularizes the toolkit and aligns it with ASP.NET Core development practices.
