+++ src/Infinity.Toolkit.AspNetCore/AuthenticationBuilderExtensions.cs
