+++ src/Infinity.Toolkit.AspNetCore/EndpointRouteBuilderExtensions.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿using Infinity.Toolkit.Handlers;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
 
-namespace Infinity.Toolkit.Handlers;
+namespace Infinity.Toolkit.AspNetCore;
 
 public static class EndpointRouteBuilderExtensions
 {
