+++ src/Infinity.Toolkit.AspNetCore/Infinity.Toolkit.AspNetCore.csproj
@@ -1,0 +1,30 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>net9.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<RootNamespace>Infinity.Toolkit.AspNetCore</RootNamespace>
+		<VersionPrefix>1.0.0</VersionPrefix>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<FrameworkReference Include="Microsoft.AspNetCore.App" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<None Include="README.md" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.9" NoWarn="NU1605" />
+		<PackageReference Include="Microsoft.Extensions.DependencyModel" Version="9.0.9" />
+		<PackageReference Include="Scrutor" Version="6.1.0" />
+		<PackageReference Include="System.Memory.Data" Version="9.0.9" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Infinity.Toolkit\Infinity.Toolkit.csproj" />
+	</ItemGroup>
+
+</Project>
