+++ src/Infinity.Toolkit.AspNetCore/README.md
@@ -1,0 +1,25 @@
+# Infinity.Toolkit.AspNetCore
+
+Infinity.Toolkit is a .NET 9.0 library that provides a set of tools and utilities for various purposes. This document provides an overview of the public methods available in the library.
+
+## Installation
+
+To install Infinity.Toolkit, add the following package reference to your project file:
+
+```<PackageReference Include="Infinity.Toolkit.AspNetCore" Version="*" />```
+
+## Usage
+
+### Public Methods
+
+#### AuthenticationBuilderExtensions
+
+- **AddMultipleBearerPolicySchemes**: Adds a policy scheme that can dynamically select the authentication scheme based on the token issuer.
+
+## Contributing
+
+Contributions are welcome! Please open an issue or submit a pull request.
+
+## License
+
+This project is licensed under the MIT License.
