--- src/Infinity.Toolkit.Handlers/Infinity.Toolkit.Handlers.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net9.0;net10.0</TargetFrameworks>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <RootNamespace>Infinity.Toolkit.Handlers</RootNamespace>
-    <VersionPrefix>1.0.0</VersionPrefix>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="README.md" Pack="true" PackagePath="" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="9.0.9" />
-    <PackageReference Include="Scrutor" Version="6.1.0" />
-    <PackageReference Include="System.Memory.Data" Version="9.0.9" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Infinity.Toolkit\Infinity.Toolkit.csproj" />
-  </ItemGroup>
-
-</Project>
