+++ src/Infinity.Toolkit.IntegrationTests/Infinity.Toolkit.IntegrationTests.csproj
@@ -10,10 +10,11 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.9" NoWarn="NU1605" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.9" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.9" />
     <PackageReference Include="Testcontainers.MsSql" Version="4.7.0" />
-    <PackageReference Include="xunit.v3.extensibility.core" Version="3.0.1" />
+    <PackageReference Include="xunit.v3.extensibility.core" Version="3.1.0" />
   </ItemGroup>
 
   <ItemGroup>
