+++ src/Infinity.Toolkit.LogFormatter/Infinity.Toolkit.LogFormatter.csproj
@@ -1,10 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net9.0;net10.0</TargetFrameworks>
+    <TargetFrameworks>net9.0;net8.0</TargetFrameworks>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <VersionPrefix>1.1.0</VersionPrefix>
+    <VersionPrefix>1.0.7</VersionPrefix>
     <PackageId>Infinity.Toolkit.LogFormatter</PackageId>
     <Description>A logging formatter that formats log messages with a Visual Studio Code inspired theme and Serilog like formatting.</Description>
     <PackageTags>Logging formatter</PackageTags>
