+++ src/Infinity.Toolkit/Handlers/HandlerContext.cs
