+++ src/Infinity.Toolkit/Handlers/IRequestHandler.cs
