+++ src/Infinity.Toolkit/Handlers/ServiceCollectionExtensions.cs
