+++ tests/Infinity.Toolkit.Tests/Infinity.Toolkit.Tests.csproj
@@ -22,6 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\Infinity.Toolkit.AspNetCore\Infinity.Toolkit.AspNetCore.csproj" />
     <ProjectReference Include="..\..\src\Infinity.Toolkit.Messaging\Infinity.Toolkit.Messaging.csproj" />
     <ProjectReference Include="..\..\src\Infinity.Toolkit.TestUtils\Infinity.Toolkit.TestUtils.csproj" />
     <ProjectReference Include="..\..\src\Infinity.Toolkit\Infinity.Toolkit.csproj" />
